pyproject.toml
[project]
name = "properapp"
version = "0.1.0"
description = ""
authors = [
    { name = "svaponi", email = "10941963+svaponi@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asgi-correlation-id (>=4.3.4,<5.0.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "asyncpg-datalayer @ git+https://github.com/svaponi/asyncpg-datalayer@v0.0.1a3",
]

[tool.poetry]
packages = [{ include = "app", from = "src" }]

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
pytest = "^9.0.2"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
testcontainers = "^4.13.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
